+++ packages/worker/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.1"></a>
+## [0.16.1](https://github.com/zetapush/zetapush/compare/v0.16.0...v0.16.1) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @zetapush/worker
+
+<a name="0.16.0"></a>
+# [0.16.0](https://github.com/zetapush/zetapush/compare/v0.15.3...v0.16.0) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @zetapush/worker
+
+<a name="0.16.0-alpha.45cf9f91"></a>
+# [0.16.0-alpha.45cf9f91](https://github.com/zetapush/zetapush/compare/v0.15.3...v0.16.0-alpha.45cf9f91) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @zetapush/worker
+
 <a name="0.15.3"></a>
 ## [0.15.3](https://github.com/zetapush/zetapush/compare/v0.15.2...v0.15.3) (2018-05-31)